chore: Remove reselect from code
Reselect seems like a really cool idea for a package. But it's all about optimizing speed, which isn't important to us during prototyping. And it will take us all time to learn how to use it effectively, and will increase the amount of time it takes for future contributors to get familiar with our codebase. So this commit removes it entirely from our code. We can re-integrate it later if we want.