feat: [WIP] Update state when there's a pending fetch of contacts
Right now, there's no piece of redux state that records whether there's a pending fetch request for contacts (e.g., when the user loads the homepage container). This commit doesn't change that, but it does lay a teeny bit of groundwork for getting there - it creates an incomplete reducer file for the homepage. We'll need to finish that file, and update the store so it listens to that reducer to update the "fetchInProgress" (or something) state property.