feat: [WIP] [broken] Server ajax call to pull data from server
This isn't working yet. The frontend is successfully getting the contacts data from the server. It's a hacky way of doing it, just so I knew we could get data from the backend properly. Now that I know we can, I'll need to work the data into the proper action/saga/reducer flow.