feat: Correctly process and display data from server
The frontend is now successfully processing and displaying the contact data it pulls from the server. When the HomePage component is loading, it fires an action to request the contacts. That is heard by the saga, which actually makes the ajax call. When the ajax call replies with data, the saga fires another action with the contacts. The reducer hears this action and loads the contacts into state.

 I had to comment out a line that filtered the stories that are displayed on each contact card's Stories dropdown. I had to get rid of that because the data we're getting from the server doesn't have a `storiesDone` property, so that line was throwing an error. Once we implement a junction table, we should rewrite the ability to filter stories for each contact.