feat: Create ability to delete a contact
The user can now click to delete a contact. Right now the feature is pretty raw - there's no confirmation dialog, and the button looks ugly and isn't probably where we want it to be. There's also no UI to let the user know a request in progress, and no way to inform the user if it didn't work (other than the contact not disappearing from the page). But it works! Those are all improvements for the future.