style: Allow user to how/hide Compose pane with new button
The last commit made a lot of improvements to the layout/appearance of the ContactCard component, but didn't actually add much functionality. Now, the button I created actually works to show/hide the Compose pane.

There are also a number of other style/layout improvements to ContactCard in this commit.