feat: [WIP] Error handling if client tries updating improper columns
When the user updates an existing contact, they should send a `values` object to the backend with the keys/columns that should be updated. This commit adds some error handling to catch if the frontend sends any keys that don't match columns in the db. In this case, the backend will send an error in the response.