feat: Ensure frontend sends correct userId to backend, pt. 2
In the previous commit, I hadn't quite finished the following: (repeat of the previous commit's message) The frontend needs to send the correct userId to the backend, so the backend can apply the requested CRUD operation to the appropriate contacts. The backend is not yet set up to handle this information. Once we have proper authentication, we won't be sending a userId; we'll be sending a jwt or something similar.

This commit should complete the frontend portion of the changes.