feat: Add "last contacted" field to Add Contact form
Previously, the Add Contact form didn't have a field for the contact's "last contacted" date. One result was that those new contacts never showed up in the dashboard.

This commit makes a number of related fixes. It adds a field in that form, and makes sure the data is handled properly by the frontend and the backend. It also makes contacts with no "last contacted" date show up in the dashboard.

This commit also includes some stylistic and small refactoring changes in files I touched while adding this functionality.