feat: Improve defaults/validation when editing/adding contacts
The Add Contact form now has a default of 180 days contact frequency. The contact's name is now also required. Ultimately it might make sense to allow a contact without a name - for example if a user gets someone's email address, but doesn't know their name - but for now, I'm going to keep it simple for me as a developer and require a name.