fix: Fix Add Contact form
I broke the Add Contact form when I consolidated it and the Edit Contact form, and I only just realized that. I was passing in an event handler with the wrong name. The way redux-forms processes the form's submit handler is a bit confusing. I had done it correctly when setting up my Edit Contact form (probably by copying from their intro example), but was getting tripped up when debugging this issue.

You pass `onSubmit` as a prop to the form. But in the form itself, you set the form's onSubmit attribute equal to the `handleSubmit` prop. But you didn't pass in a handleSubmit prop. Apparently when you pass `onSubmit` to the higher-order redux-form component, it converts that into a `handleSubmit` prop, which is passed on down.

It's working now. 'nuff said.