feat: Add "Today" button to Last Contacted form field
A minor, but possibly frequent, inconvenience while adding a new contact (or editing a contact) was that if the user had just contacted the contact today, they would frequently have to type in today's date. The date picker is a little more annoying than other inputs to fill out, so this could get old fast.

Since it's likely to be a pretty frequent thing to want to do, I added a "Today" button next to the form field. Clicking that button updates the field with today's date.