feat: Normalize/mask phone number; fix validation
This commit adds phone number masking. I.e., when the user types in numbers, they see the formatting added for them. So they never have to type the parens, spaces or hyphens; those are provided for them. The placeholder text also shows them the format their phone number should go in. The idea is that they have the convenience of never needing to type anything other than numbers, and there is also no ambiguity about what sorts of phone number formats are "allowed" by our app, since we provide the format.

A downside is that right now, we're not supporting any format other than the standard US 10-digit phone number. Ultimately we'll need to become more flexible.

This commit also:
 - updates the relevant validation functions (now allowing non-numeric characters like parens)
 - updates how we store dates in state and in the db. Now dates, when they exist, are always stored in "YYYY-MM-DD" format, rather than storing them in longer strings in the DB, and storing them as `moment` objects in state. I was running into complications because sometimes dates were moments, and sometimes they were strings, and I was beginning to run into errors. Now we'll only use `moment` to help us make comparisons or get a certain formatting, but we won't store things as moments.